Refactor repo structure into `components`
