+++ rust-toolchain.toml
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "1.68.2"
+channel = "1.69"
 components = ["clippy", "rustfmt"]
