+++ src/components/stats.rs
