+++ src/draw.rs
@@ -6,7 +6,7 @@ use tui::widgets::{Block, BorderType, Borders, Clear, Paragraph, Tabs};
 use tui::{Frame, Terminal};
 
 use crate::app::{App, DebugState, MenuItem};
-use crate::debug::DebugInfo;
+use crate::components::debug::DebugInfo;
 use crate::ui::at_bat::AtBatWidget;
 use crate::ui::boxscore::TeamBatterBoxscoreWidget;
 use crate::ui::help::{HelpWidget, DOCS_LEN};
