+++ src/ui/stats.rs
@@ -1,6 +1,8 @@
 use mlb_api::client::StatGroup;
 
-use crate::stats::{StatsState, TeamOrPlayer, STATS_DEFAULT_COL_WIDTH, STATS_FIRST_COL_WIDTH};
+use crate::components::stats::{
+    StatsState, TeamOrPlayer, STATS_DEFAULT_COL_WIDTH, STATS_FIRST_COL_WIDTH,
+};
 
 use tui::{
     buffer::Buffer,
